Add response of the /me endpoint to /login endpoint
