+++ packages/rocketchat-api/server/api.js
@@ -60,6 +60,10 @@ class API extends Restivus {
 		return RocketChat.API.helperMethods;
 	}
 
+	getHelperMethod(name) {
+		return RocketChat.API.helperMethods.get(name);
+	}
+
 	addAuthMethod(method) {
 		this.authMethods.push(method);
 	}
@@ -224,6 +228,7 @@ class API extends Restivus {
 		this.addRoute('login', {authRequired: false}, {
 			post() {
 				const args = loginCompatibility(this.bodyParams);
+				const getUserInfo = self.getHelperMethod('getUserInfo');
 
 				const invocation = new DDPCommon.MethodInvocation({
 					connection: {
@@ -273,7 +278,8 @@ class API extends Restivus {
 					status: 'success',
 					data: {
 						userId: this.userId,
-						authToken: auth.token
+						authToken: auth.token,
+						me: getUserInfo(this.userId)
 					}
 				};
